refactoring the route
